pyproject.toml
[project]
name = "is1ab-ctf-template"
version = "1.0.0"
description = "IS1AB CTF Challenge Development Template"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
